Use Microsoft.Extensions.Options for adapter configuration
- Removes `IAdapterOptionsMonitor` and `AdapterOptionsMonitor` types.
- Adds `Microsoft.Extensions.Options` to allow adapter options to be passed to adapters using `IOptions<T>`, `IOptionsSnapshot<T>` or `IOptionsMonitor<T>`.
- Refactors `AdapterBase` and `AdapterBase<T>` so that the latter is now the base class of the former (instead of the other way around).
- Updates ASP.NET Core sample app to demonstrate how adapter options can be specified in application configuration (e.g. in `appsettings.json`).
