+++ examples/DataCore.Adapter.AspNetCoreExample/appsettings.json
@@ -9,5 +9,13 @@
     "EndpointDefaults": {
       "Protocols": "Http1AndHttp2"
     }
+  },
+  "CsvAdapter": {
+    "sensor-csv": {
+      "Name": "Sensor CSV",
+      "Description": "CSV adapter with dummy sensor data",
+      "IsDataLoopingAllowed": true,
+      "CsvFile": "DummySensorData.csv"
+    }
   }
 }
