+++ src/DataCore.Adapter/AdapterConstants.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DataCore.Adapter {
+
+    /// <summary>
+    /// Adapter-related constant values.
+    /// </summary>
+    public static class AdapterConstants {
+
+        /// <summary>
+        /// Maximum length for an adapter ID.
+        /// </summary>
+        public const int MaxIdLength = 100;
+
+        /// <summary>
+        /// Maximum length for an adapter name.
+        /// </summary>
+        public const int MaxNameLength = 200;
+
+        /// <summary>
+        /// Maximum length for an adapter description.
+        /// </summary>
+        public const int MaxDescriptionLength = 500;
+
+    }
+}
