+++ src/DataCore.Adapter/README.md
@@ -2,7 +2,7 @@
 
 Contains base classes and utility classes for implementing an [IAdapter](/src/DataCore.Adapter.Abstractions/IAdapter.cs).
 
-Extend from [AdapterBase](./AdapterBase.cs) for easy implementation.
+Extend from [AdapterBase](./AdapterBase.cs) for easy implementation or [AdapterBase<T>](./AdapterBaseT.cs) if you need to supply configurable options to your adapter.
 
 
 # Installation
@@ -232,5 +232,120 @@ When writing an extension feature, methods can be annotated with an [ExtensionFe
     "description": "The pong message.",
     "exampleValue": "{ \"CorrelationId\": \"10085252-5b69-4ca2-a727-c850c7825630\" }"
   }
+}
+```
+
+
+# Providing Adapter Options From Configuration
+
+`AdapterBase<T>` and `AdapterBase` both define constructors that allow the options for the adapter to be supplied via an `IOptions<T>`, `IOptionsSnapshot<T>` or `IOptionsMonitor<T>` instance supplied by the configuration system of an ASP.NET Core application, or an application using the .NET Core Generic Host.
+
+When supplying an `IOptionsSnapshot<T>` or `IOptionsMonitor<T>`, the adapter ID will be used as the key when retrieving the options instance. If an entry is not defined, the default entry (i.e. the entry with a key defined by `Microsoft.Extensions.Options.Options.DefaultName`) will be used, if available.
+
+For example:
+
+```json
+// appsettings.json
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+  "Kestrel": {
+    "EndpointDefaults": {
+      "Protocols": "Http1AndHttp2"
+    }
+  },
+  "CsvAdapter": {
+    "my-csv": {
+      "Name": "Sample Data",
+      "Description": "CSV adapter with dummy data",
+      "IsDataLoopingAllowed": true,
+      "CsvFile": "SampleData.csv"
+    }
+  }
+}
+```
+
+```csharp
+public class Startup {
+
+    public IConfiguration Configuration { get; }
+
+    public Startup(IConfiguration configuration) {
+        Configuration = configuration;
+    }
+
+    public void ConfigureServices(IServiceCollection services) {
+        // Other configuration removed for brevity.
+
+        // Bind CSV adapter options against the application configuration.
+        services.Configure<DataCore.Adapter.Csv.CsvAdapterOptions>(Configuration.GetSection("CsvAdapter:my-csv"));
+
+        services
+            .AddDataCoreAdapterAspNetCoreServices()
+            .AddHostInfo(HostInfo.Create(
+                "My Host",
+                "A brief description of the hosting application",
+                "0.9.0-alpha", // SemVer v2
+                VendorInfo.Create("Intelligent Plant", "https://appstore.intelligentplant.com"),
+                AdapterProperty.Create("Project URL", "https://github.com/intelligentplant/AppStoreConnect.Adapters")
+            ))
+            // Create adapter using an IOptions<T> to supply options.
+            .AddAdapter<DataCore.Adapter.Csv.CsvAdapter>(sp => ActivatorUtilities.CreateInstance<Csv.CsvAdapter>(
+                sp, 
+                "my-csv", // Adapter ID 
+                sp.GetRequiredService<IOptions<DataCore.Adapter.Csv.CsvAdapterOptions>>()
+            ))
+            .AddAdapterFeatureAuthorization<MyAdapterFeatureAuthHandler>();
+    }
+
+    // Remaining code removed for brevity.
+
+}
+```
+
+Note that, when using `IOptionsSnapshot<T>` or `IOptionsMonitor<T>`, the adapter will always try and retrieve named options that match the ID of the adapter. That is, if you register an adapter with an ID of `adapter-001`, you must also register named options with the configuration system with a name of `adapter-001`:
+
+```csharp
+public class Startup {
+
+    public IConfiguration Configuration { get; }
+
+    public Startup(IConfiguration configuration) {
+        Configuration = configuration;
+    }
+
+    public void ConfigureServices(IServiceCollection services) {
+        // Other configuration removed for brevity.
+
+        // Bind named CSV adapter options against the application configuration.
+        services.Configure<DataCore.Adapter.Csv.CsvAdapterOptions>(
+            "my-csv", // Key for this set of options
+            Configuration.GetSection("CsvAdapter:my-csv")
+        );
+
+        services
+            .AddDataCoreAdapterAspNetCoreServices()
+            .AddHostInfo(HostInfo.Create(
+                "My Host",
+                "A brief description of the hosting application",
+                "0.9.0-alpha", // SemVer v2
+                VendorInfo.Create("Intelligent Plant", "https://appstore.intelligentplant.com"),
+                AdapterProperty.Create("Project URL", "https://github.com/intelligentplant/AppStoreConnect.Adapters")
+            ))
+            // Create adapter using an IOptionsSnapshot<T> to supply named options.
+            .AddAdapter<DataCore.Adapter.Csv.CsvAdapter>(sp => ActivatorUtilities.CreateInstance<Csv.CsvAdapter>(
+                sp, 
+                "my-csv", // Adapter ID; also used as the named options key   
+                sp.GetRequiredService<IOptionsSnapshot<DataCore.Adapter.Csv.CsvAdapterOptions>>()
+            ))
+            .AddAdapterFeatureAuthorization<MyAdapterFeatureAuthHandler>();
+    }
+
+    // Remaining code removed for brevity.
+
 }
 ```
